Organize files under src folder
