+++ src/pages/Signup.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { supabase } from '../supabase';
+import { supabase } from '../supabase.js';
 import { useNavigate } from 'react-router-dom';
 
 export default function Signup() {
